[IMP] runbot: also use module name when fingerprinting
When build logs are scanned, only the message is used for the
fingerprint. When a message is too generic, the created build error use
the first module name it finds but links all other builds with the same
message to this build error.

It's preferable to create more build errors when they occur in different
modules. With this commit, the module name is taken into account when
computing the fingerprint.