[IMP] runbot_merge: optimise edited event
For the longest time Github's `change` key was borked when
transitioning a description to and from empty. They fixed that during
2023, which I already saw and impacted on
DC (xmo-odoo/dummy_central@1ebed9d4183d34fa412f6556aedde019803a15a0
and xmo-odoo/dummy_central@937e87c2a4383f5ffcda6756443b082d2bde14bd)
but this had yet to be taken in account by the mergebot.

This is now done, the code is functionally reverted to what it was
before I realised `changes` was hosed and moved off of it in
3da18741963e80b34fd98055ecb6bc1483d09c35.

Fixes #743