[FIX] runbot: add missing repos
On prepare, if a commit is pushed, it will be added to the bundle
if no trigger exist in the project, or if a repo is not used by any
trigger, it may appear in the batch from when pushed but not in other
case, creating a strange behaviour. This is mainly the case when a repo
is in a project without tirgger, just a group of commit.