[IMP] runbot_merge: a PR can't be reopened if its batch is merged
In that case, ignore the reopen, close the PR, and tell the idiot to
fuck off.

Also the case where a PR is reopened while its batch is staged was
already handled, but probably not tested: it was implicit in
forcefully updating the HEAD of the PR, which triggers an unstage
since c8a06601a7c223cdfc9da23f2c799b4c660f274b.

Now that scenario is tested, which should lower the odds of breaking
it in the future.

Fixes #965